refactor a bunch of OutputUtils stuff that really belongs in OutputToApiDataSideOutput adapter
